Rename compiler.js to compile-harmony.js
Again, using the same sort of naming as the less and coffee script
packages.
