+++ plugin/compile-harmony.js
